Move setSelection to selection.set.
