+++ src/components/Editable.tsx
@@ -56,7 +56,6 @@ import {
   isHTML,
   isURL,
   pathToContext,
-  setSelection,
   strip,
   normalizeThought,
 } from '../util'
@@ -277,7 +276,7 @@ const Editable = ({
       setCursor({
         cursorHistoryClear: true,
         editing,
-        // set offset to null to prevent setSelection on next render
+        // set offset to null to prevent selection.set on next render
         // to use the existing offset after a user clicks or touches the screent
         // when cursor is changed through another method, such as cursorDown, offset will be reset
         offset: null,
@@ -362,7 +361,7 @@ const Editable = ({
 
   /** Set the selection to the current Editable at the cursor offset. */
   const setSelectionToCursorOffset = () => {
-    setSelection(contentRef.current, { offset: cursorOffset || state.cursorOffset || 0 })
+    selection.set(contentRef.current, { offset: cursorOffset || state.cursorOffset || 0 })
   }
 
   useEffect(() => {
@@ -381,7 +380,7 @@ const Editable = ({
     // focus on the ContentEditable element if editing os on desktop
     const editMode = !isTouch || editing
 
-    // if there is no browser selection, do not manually call setSelection as it does not preserve the cursor offset. Instead allow the default focus event.
+    // if there is no browser selection, do not manually call selection.set as it does not preserve the cursor offset. Instead allow the default focus event.
     const cursorWithoutSelection = state.cursorOffset !== null || !selection.isActive()
 
     // if the selection is at the beginning of the thought, ignore cursorWithoutSelection and allow the selection to be set
@@ -390,7 +389,7 @@ const Editable = ({
     const isAtBeginning = !isTouch && selection.offset() === 0
 
     /**
-     * Note: There are a lot of different values that determine if setSelection is called!
+     * Note: There are a lot of different values that determine if we set the selection
      * You may need to inspect them if something goes wrong.
      */
     // if (isEditing) {
