+++ src/components/Note.tsx
@@ -4,7 +4,7 @@ import { isTouch } from '../browser'
 import { store } from '../store'
 import { attribute, getParent, isContextViewActive } from '../selectors'
 import { deleteAttribute, editing, setAttribute, setNoteFocus } from '../action-creators'
-import { setSelection, strip } from '../util'
+import { strip } from '../util'
 import ContentEditable, { ContentEditableEvent } from 'react-contenteditable'
 import asyncFocus from '../device/asyncFocus'
 import selectNextEditable from '../device/selectNextEditable'
@@ -51,7 +51,7 @@ const Note = ({ context, onFocus }: NoteProps) => {
       e.stopPropagation()
       e.preventDefault()
       editable.focus()
-      setSelection(editable, { end: true })
+      selection.set(editable, { end: true })
       dispatch(setNoteFocus({ value: false }))
     }
     // delete empty note
@@ -65,7 +65,7 @@ const Note = ({ context, onFocus }: NoteProps) => {
         asyncFocus()
       }
       editable.focus()
-      setSelection(editable, { end: true })
+      selection.set(editable, { end: true })
 
       dispatch(deleteAttribute({ context, key: '=note' }))
       dispatch(setNoteFocus({ value: false }))
