+++ src/shortcuts/archive.native.tsx
@@ -1,11 +1,11 @@
 import React from 'react'
 import { Key } from 'ts-key-enum'
-
 import { hasChild } from '../selectors'
-import { ellipsize, headValue, isEM, isRoot, pathToContext, setSelection } from '../util'
+import { ellipsize, headValue, isEM, isRoot, pathToContext } from '../util'
 import { alert, archiveThought, deleteAttribute, error } from '../action-creators'
 import { Icon as IconType, Shortcut } from '../@types'
 import Svg, { G, Path } from 'react-native-svg'
+import * as selection from '../device/selection'
 
 let undoArchiveTimer: number // eslint-disable-line fp/no-let
 
@@ -34,7 +34,7 @@ const exec: Shortcut['exec'] = (dispatch, getState, e) => {
 
       if (editable) {
         editable.focus()
-        setSelection(editable, { end: true })
+        selection.set(editable, { end: true })
       }
     } else {
       // clear the undo alert timer to prevent previously cleared undo alert from closing this one
