+++ src/shortcuts/archive.tsx
@@ -2,9 +2,10 @@ import React from 'react'
 import { Key } from 'ts-key-enum'
 import { isTouch } from '../browser'
 import { hasChild } from '../selectors'
-import { ellipsize, isDocumentEditable, headValue, isEM, isRoot, pathToContext, setSelection } from '../util'
+import { ellipsize, isDocumentEditable, headValue, isEM, isRoot, pathToContext } from '../util'
 import { alert, archiveThought, deleteAttribute, error } from '../action-creators'
 import asyncFocus from '../device/asyncFocus'
+import * as selection from '../device/selection'
 import { Icon as IconType, Shortcut } from '../@types'
 
 let undoArchiveTimer: number // eslint-disable-line fp/no-let
@@ -35,7 +36,7 @@ const exec: Shortcut['exec'] = (dispatch, getState, e) => {
       }
       if (editable) {
         editable.focus()
-        setSelection(editable, { end: true })
+        selection.set(editable, { end: true })
       }
     } else {
       // clear the undo alert timer to prevent previously cleared undo alert from closing this one
