+++ src/shortcuts/note.native.ts
@@ -1,10 +1,11 @@
 import { attribute, hasChild, simplifyPath } from '../selectors'
 
 import PencilIcon from '../components/icons/PencilIcon'
-import { editableNode, pathToContext, setSelection } from '../util'
+import { editableNode, pathToContext } from '../util'
 import { setAttribute, setNoteFocus } from '../action-creators'
 import { Shortcut } from '../@types'
 import { HOME_PATH } from '../constants'
+import * as selection from '../device/selection'
 
 const noteShortcut: Shortcut = {
   id: 'note',
@@ -44,7 +45,7 @@ const noteShortcut: Shortcut = {
         // select thought
         if (noteFocus) {
           thoughtEl.focus()
-          setSelection(thoughtEl, { end: true })
+          selection.set(thoughtEl, { end: true })
           dispatch(setNoteFocus({ value: false }))
         }
         // select note
@@ -55,7 +56,7 @@ const noteShortcut: Shortcut = {
           const noteEl = closest.querySelector('.note [contenteditable]') as HTMLElement
           if (!noteEl) return
           noteEl.focus()
-          setSelection(noteEl, { end: true })
+          selection.set(noteEl, { end: true })
         }
       } catch (e) {
         console.warn('Note element not found in DOM.', context)
