+++ src/shortcuts/note.ts
@@ -1,11 +1,12 @@
 import { isTouch } from '../browser'
 import { attribute, hasChild, simplifyPath } from '../selectors'
 import PencilIcon from '../components/icons/PencilIcon'
-import { editableNode, isDocumentEditable, pathToContext, setSelection } from '../util'
+import { editableNode, isDocumentEditable, pathToContext } from '../util'
 import { setAttribute, setNoteFocus } from '../action-creators'
 import asyncFocus from '../device/asyncFocus'
 import { Shortcut } from '../@types'
 import { HOME_PATH } from '../constants'
+import * as selection from '../device/selection'
 
 const noteShortcut: Shortcut = {
   id: 'note',
@@ -50,7 +51,7 @@ const noteShortcut: Shortcut = {
         // select thought
         if (noteFocus) {
           thoughtEl.focus()
-          setSelection(thoughtEl, { end: true })
+          selection.set(thoughtEl, { end: true })
           dispatch(setNoteFocus({ value: false }))
         }
         // select note
@@ -60,7 +61,7 @@ const noteShortcut: Shortcut = {
           const noteEl = closest.querySelector('.note [contenteditable]') as HTMLElement
           if (!noteEl) return
           noteEl.focus()
-          setSelection(noteEl, { end: true })
+          selection.set(noteEl, { end: true })
         }
       } catch (e) {
         console.warn('Note element not found in DOM.', context)
