+++ src/util/index.ts
@@ -82,7 +82,6 @@ export { removeContext } from './removeContext'
 export { removeDuplicatedContext } from './removeDuplicatedContext'
 export { removeHome } from './removeHome'
 export { roamJsonToBlocks } from './roamJsonToBlocks'
-export { setSelection } from './setSelection'
 export { sort } from './sort'
 export { spellNumber } from './spellNumber'
 export { splice } from './splice'
